layer/requirements.txt
bcrypt==3.1.7
boto3==1.9.250
botocore==1.12.250
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
cryptography==2.8
docutils==0.15.2
idna==2.8
jmespath==0.9.4
paramiko==2.6.0
pycparser==2.19
PyNaCl==1.3.0
python-dateutil==2.8.0
PyYAML==5.1.2
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
spur==0.3.21
termcolor==1.1.0
urllib3==1.25.6
